Crate edit

Modules

apperr

Provides a transparent error type for edit.

arena

Arena allocators. Small and fast.

base64

Base64 facilities.

buffer

A text buffer for a text editor.

cell

std::cell::RefCell, but without runtime checks in release builds.

clipboard

Clipboard facilities for the editor.

document

Abstractions over reading/writing arbitrary text containers.

framebuffer

A shoddy framebuffer for terminal applications.

fuzzy

Fuzzy search algorithm based on the one used in VS Code (/src/vs/base/common/fuzzyScorer.ts). Other algorithms exist, such as Sublime Text’s, or the one used in fzf, but I figured that this one is what lots of people may be familiar with.

hash

Provides fast, non-cryptographic hash functions.

helpers

Random assortment of helpers I didn’t know where to put.

icu

Bindings to the ICU library.

input

Parses VT sequences into input events.

oklab

Oklab colorspace conversions.

path

Path related helpers.

simd

Provides various high-throughput utilities.

sys

Platform abstractions.

tui

An immediate mode UI framework for terminals.

unicode

Everything related to Unicode lives here.

vt

Our VT parser.

Macros

arena_format