edit::buffer

Struct GapBuffer

pub struct GapBuffer { /* private fields */ }

Most people know how Vec<T> works: It has some spare capacity at the end, so that pushing into it doesn’t reallocate every single time. A gap buffer is the same thing, but the spare capacity can be anywhere in the buffer. This variant is optimized for large buffers and uses virtual memory.

Implementations

impl GapBuffer

pub fn new(small: bool) -> Result<Self>

pub fn len(&self) -> usize

pub fn generation(&self) -> u32

pub fn set_generation(&mut self, generation: u32)

pub fn allocate_gap( &mut self, off: usize, len: usize, delete: usize, ) -> &mut [u8]

WARNING: The returned slice must not necessarily be the same length as len (due to OOM).

pub fn commit_gap(&mut self, len: usize)

pub fn replace(&mut self, range: Range<usize>, src: &[u8])

pub fn clear(&mut self)

pub fn extract_raw( &self, range: Range<usize>, out: &mut Vec<u8>, out_off: usize, )

pub fn copy_from(&mut self, src: &dyn ReadableDocument) -> bool

Replaces the entire buffer contents with the given text. The method is optimized for the case where the given text already matches the existing contents. Returns true if the buffer contents were changed.

pub fn copy_into(&self, dst: &mut dyn WriteableDocument)

Copies the contents of the buffer into a string.