edit::icu

Struct Converter

pub struct Converter<'pivot> { /* private fields */ }

Converts between two encodings using ICU.

Implementations

impl<'pivot> Converter<'pivot>

pub fn new( pivot_buffer: &'pivot mut [MaybeUninit<u16>], source_encoding: &str, target_encoding: &str, ) -> Result<Self>

Constructs a new Converter instance.

Parameters

pivot_buffer: A buffer used to cache partial conversions. Don’t make it too small.

source_encoding: The source encoding name (e.g., “UTF-8”).

target_encoding: The target encoding name (e.g., “UTF-16”).

pub fn convert( &mut self, input: &[u8], output: &mut [MaybeUninit<u8>], ) -> Result<(usize, usize)>

Performs one step of the encoding conversion.

Parameters

input: The input buffer to convert from. It should be in the source_encoding that was previously specified.

output: The output buffer to convert to. It should be in the target_encoding that was previously specified.

Returns

A tuple containing:

The number of bytes read from the input buffer.

The number of bytes written to the output buffer.