edit::simd::lines_bwd

Function lines_bwd

pub fn lines_bwd(
    haystack: &[u8],
    offset: usize,
    line: CoordType,
    line_stop: CoordType,
) -> (usize, CoordType)

Starting from the offset in haystack with a current line index of line, this seeks backwards to the line_stop-nth line and returns the new offset and the line index at that point.

Note that this function differs from lines_fwd in that it seeks backwards even if the line is already at line_stop. This allows you to ensure (or test) whether offset is at a line start.

It returns an offset past a newline and thus at the start of a line.