edit

Module sys

Platform abstractions.

Structs

Deinit

FileId

LibIcu

Functions

apperr_format

apperr_is_not_found

canonicalize

Returns the canonical, absolute form of a path with all intermediate components normalized and symbolic links resolved.

file_id

Returns a unique identifier for the given file by handle or path.

get_proc_address⚠

Loads a function from a dynamic library.

icu_add_renaming_suffix

icu_detect_renaming_suffix

ICU, by default, adds the major version as a suffix to each exported symbol. They also recommend to disable this for system-level installations (runConfigureICU Linux --disable-renaming), but I found that many (most?) Linux distributions don’t do this for some reason. This function returns the suffix, if any.

init

inject_window_size_into_stdin

load_icu

open_stdin_if_redirected

preferred_languages

read_stdin

Reads from stdin.

switch_modes

virtual_commit⚠

Commits a virtual memory region of the given size.

virtual_release⚠

Releases a virtual memory region of the given size.

virtual_reserve⚠

Reserves a virtual memory region of the given size. To commit the memory, use virtual_commit. To release the memory, use virtual_release.

write_stdout