edit::tui

Struct Context

pub struct Context<'a, 'input> { /* private fields */ }

Context is a temporary object that is created for each frame. Its primary purpose is to build a UI tree.

Implementations

impl<'a> Context<'a, '_>

pub fn arena(&self) -> &'a Arena

Get an arena for temporary allocations such as for arena_format.

pub fn size(&self) -> Size

Returns the viewport size.

pub fn indexed(&self, index: IndexedColor) -> StraightRgba

Returns an indexed color from the framebuffer.

pub fn indexed_alpha( &self, index: IndexedColor, numerator: u32, denominator: u32, ) -> StraightRgba

Returns an indexed color from the framebuffer with the given alpha. See Framebuffer::indexed_alpha().

pub fn contrasted(&self, color: StraightRgba) -> StraightRgba

Returns a color in contrast with the given color. See Framebuffer::contrasted().

pub fn clipboard_ref(&self) -> &Clipboard

Returns the clipboard.

pub fn clipboard_mut(&mut self) -> &mut Clipboard

Returns the clipboard (mutable).

pub fn needs_rerender(&mut self)

Tell the UI framework that your state changed and you need another layout pass.

pub fn block_begin(&mut self, classname: &'static str)

Begins a generic UI block (container) with a unique ID derived from the given classname.

pub fn block_end(&mut self)

Ends the current UI block, returning to its parent container.

pub fn next_block_id_mixin(&mut self, id: u64)

Mixes in an extra value to the next UI block’s ID for uniqueness. Use this when you build a list of items with the same classname.

pub fn focus_on_first_present(&mut self)

If this is the first time the current node is being drawn, it’ll steal the active focus.

pub fn steal_focus(&mut self)

Steals the focus unconditionally.

pub fn toss_focus_up(&mut self)

If the current node owns the focus, it’ll be given to the parent.

pub fn inherit_focus(&mut self)

If the parent node owns the focus, it’ll be given to the current node.

pub fn attr_focus_well(&mut self)

Causes keyboard focus to be unable to escape this node and its children. It’s a “well” because if the focus is inside it, it can’t escape.

pub fn attr_intrinsic_size(&mut self, size: Size)

Explicitly sets the intrinsic size of the current node. The intrinsic size is the size the node ideally wants to be.

pub fn attr_float(&mut self, spec: FloatSpec)

Turns the current node into a floating node, like a popup, modal or a tooltip.

pub fn attr_border(&mut self)

Gives the current node a border.

pub fn attr_position(&mut self, align: Position)

Sets the current node’s position inside the parent.

pub fn attr_padding(&mut self, padding: Rect)

Assigns padding to the current node.

pub fn attr_background_rgba(&mut self, bg: StraightRgba)

Assigns a sRGB background color to the current node.

pub fn attr_foreground_rgba(&mut self, fg: StraightRgba)

Assigns a sRGB foreground color to the current node.

pub fn attr_reverse(&mut self)

Applies reverse-video to the current node: Background and foreground colors are swapped.

pub fn consume_shortcut(&mut self, shortcut: InputKey) -> bool

Checks if the current keyboard input matches the given shortcut, consumes it if it is and returns true in that case.

pub fn keyboard_input(&self) -> Option<InputKey>

Returns current keyboard input, if any. Returns None if the input was already consumed.

pub fn set_input_consumed(&mut self)

pub fn was_mouse_down(&mut self) -> bool

Returns whether the mouse was pressed down on the current node.

pub fn contains_mouse_down(&mut self) -> bool

Returns whether the mouse was pressed down on the current node’s subtree.

pub fn is_focused(&mut self) -> bool

Returns whether the current node is focused.

pub fn contains_focus(&mut self) -> bool

Returns whether the current node’s subtree is focused.

pub fn modal_begin(&mut self, classname: &'static str, title: &str)

Begins a modal window. Call Context::modal_end().

pub fn modal_end(&mut self) -> bool

Ends the current modal window block. Returns true if the user pressed Escape (a request to close).

pub fn table_begin(&mut self, classname: &'static str)

Begins a table block. Call Context::table_end(). Tables are the primary way to create a grid layout, and to layout controls on a single row (= a table with 1 row).

pub fn table_set_columns(&mut self, columns: &[CoordType])

Assigns widths to the columns of the current table. By default, the table will left-align all columns.

pub fn table_set_cell_gap(&mut self, cell_gap: Size)

Assigns the gap between cells in the current table.

pub fn table_next_row(&mut self)

Starts the next row in the current table.

pub fn table_end(&mut self)

Ends the current table block.

pub fn label(&mut self, classname: &'static str, text: &str)

Creates a simple text label.

pub fn styled_label_begin(&mut self, classname: &'static str)

Creates a styled text label.

Example

use edit::framebuffer::IndexedColor;
use edit::tui::Context;

fn draw(ctx: &mut Context) {
    ctx.styled_label_begin("label");
    // Shows "Hello" in the inherited foreground color.
    ctx.styled_label_add_text("Hello");
    // Shows ", World!" next to "Hello" in red.
    ctx.styled_label_set_foreground(ctx.indexed(IndexedColor::Red));
    ctx.styled_label_add_text(", World!");
}

pub fn styled_label_set_foreground(&mut self, fg: StraightRgba)

Changes the active pencil color of the current label.

pub fn styled_label_set_attributes(&mut self, attr: Attributes)

Changes the active pencil attributes of the current label.

pub fn styled_label_add_text(&mut self, text: &str)

Adds text to the current label.

pub fn styled_label_end(&mut self)

Ends the current label block.

pub fn attr_overflow(&mut self, overflow: Overflow)

Sets the overflow behavior of the current label.

pub fn button( &mut self, classname: &'static str, text: &str, style: ButtonStyle, ) -> bool

Creates a button with the given text. Returns true if the button was activated.

pub fn checkbox( &mut self, classname: &'static str, text: &str, checked: &mut bool, ) -> bool

Creates a checkbox with the given text. Returns true if the checkbox was activated.

pub fn editline( &mut self, classname: &'static str, text: &mut dyn WriteableDocument, ) -> bool

Creates a text input field. Returns true if the text contents changed.

pub fn textarea(&mut self, classname: &'static str, tb: RcTextBuffer)

Creates a text area.

pub fn scrollarea_begin( &mut self, classname: &'static str, intrinsic_size: Size, )

Creates a scrollable area.

pub fn scrollarea_scroll_to(&mut self, pos: Point)

Scrolls the current scrollable area to the given position.

pub fn scrollarea_end(&mut self)

Ends the current scrollarea block.

pub fn list_begin(&mut self, classname: &'static str)

Creates a list where exactly one item is selected.

pub fn list_item(&mut self, select: bool, text: &str) -> ListSelection

Creates a list item with the given text.

pub fn styled_list_item_begin(&mut self)

Creates a list item consisting of a styled label. See Context::styled_label_begin.

pub fn styled_list_item_end(&mut self, select: bool) -> ListSelection

Ends the current styled list item.

pub fn list_item_steal_focus(&mut self)

Context::steal_focus, but for a list view.

This exists, because didn’t want to figure out how to get Context::styled_list_item_end to recognize a regular, programmatic focus steal.

pub fn list_end(&mut self)

Ends the current list block.

pub fn menubar_begin(&mut self)

Creates a menubar, to be shown at the top of the screen.

pub fn menubar_menu_begin(&mut self, text: &str, accelerator: char) -> bool

Appends a menu to the current menubar.

Returns true if the menu is open. Continue appending items to it in that case.

pub fn menubar_menu_button( &mut self, text: &str, accelerator: char, shortcut: InputKey, ) -> bool

Appends a button to the current menu.

pub fn menubar_menu_checkbox( &mut self, text: &str, accelerator: char, shortcut: InputKey, checked: bool, ) -> bool

Appends a checkbox to the current menu. Returns true if the checkbox was activated.

pub fn menubar_menu_end(&mut self)

Ends the current menu.

pub fn menubar_end(&mut self)

Ends the current menubar.