edit::tui

Struct Tui

pub struct Tui { /* private fields */ }

There’s two types of lifetimes the TUI code needs to manage:

Across frames

Per frame

Tui manages the first one. It’s also the entrypoint for everything else you may want to do.

Implementations

impl Tui

pub fn new() -> Result<Self>

Creates a new Tui instance for storing state across frames.

pub fn setup_indexed_colors(&mut self, colors: [StraightRgba; 18])

Sets up the framebuffer’s color palette.

pub fn setup_modifier_translations( &mut self, translations: ModifierTranslations, )

Set up translations for Ctrl/Alt/Shift modifiers.

pub fn set_floater_default_bg(&mut self, color: StraightRgba)

Set the default background color for floaters (dropdowns, etc.).

pub fn set_floater_default_fg(&mut self, color: StraightRgba)

Set the default foreground color for floaters (dropdowns, etc.).

pub fn set_modal_default_bg(&mut self, color: StraightRgba)

Set the default background color for modals.

pub fn set_modal_default_fg(&mut self, color: StraightRgba)

Set the default foreground color for modals.

pub fn read_timeout(&mut self) -> Duration

If the TUI is currently running animations, etc., this will return a timeout smaller than time::Duration::MAX.

pub fn size(&self) -> Size

Returns the viewport size.

pub fn indexed(&self, index: IndexedColor) -> StraightRgba

Returns an indexed color from the framebuffer.

pub fn indexed_alpha( &self, index: IndexedColor, numerator: u32, denominator: u32, ) -> StraightRgba

Returns an indexed color from the framebuffer with the given alpha. See Framebuffer::indexed_alpha().

pub fn contrasted(&self, color: StraightRgba) -> StraightRgba

Returns a color in contrast with the given color. See Framebuffer::contrasted().

pub fn clipboard_ref(&self) -> &Clipboard

Returns the clipboard.

pub fn clipboard_mut(&mut self) -> &mut Clipboard

Returns the clipboard (mutable).

pub fn create_context<'a, 'input>( &'a mut self, input: Option<Input<'input>>, ) -> Context<'a, 'input>

Starts a new frame and returns a Context for it.

pub fn needs_settling(&mut self) -> bool

After you finished processing all input, continue redrawing your UI until this returns false.

pub fn render<'a>(&mut self, arena: &'a Arena) -> ArenaString<'a>

Renders the last frame into the framebuffer and returns the VT output.

pub fn debug_layout<'a>(&mut self, arena: &'a Arena) -> ArenaString<'a>

Outputs a debug string of the layout and focus tree.